separate query building from query fetching
